+++ nlubridge/dataloaders/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2021 Klaus-Peter Engelbrecht, Deutsche Telekom AG
+# This software is distributed under the terms of the MIT license
+# which is available at https://opensource.org/licenses/MIT
+"""Provides a unified API to several training data formats."""  # noqa D400
