+++ tests/fixtures/rasa3_nlu.yml
@@ -1,4 +1,3 @@
-version: "3.0"
 nlu:
 - intent: restaurant_search
   examples: |
