+++ tests/test_dataset_loaders.py
@@ -1,5 +1,4 @@
 import os
-from packaging import version
 from test_datasets import FIXTURE_PATH
 
 
@@ -9,9 +8,9 @@ def test_from_huggingface():
 
 
 def test_from_watson():
-    from nlubridge.vendors.watson import load_data
+    from nlubridge import from_watson
 
-    ds = load_data(os.path.join(FIXTURE_PATH, "watson_intents_export.csv"))
+    ds = from_watson(os.path.join(FIXTURE_PATH, "watson_intents_export.csv"))
     assert len(ds) == 4
     assert ds.texts[0][0] != '"'
     assert ds.texts[2] == "testing umläuts"
@@ -20,14 +19,9 @@ def test_from_watson():
 
 
 def test_from_rasa_json():
-    from rasa import __version__ as rasa_version
+    from nlubridge import from_rasa
 
-    if version.parse(rasa_version) >= version.parse("3.0.0"):
-        return  # see 'test_from_rasa3_yml' for Rasa 3.x
-
-    from nlubridge.vendors.rasa import load_data
-
-    ds = load_data(os.path.join(FIXTURE_PATH, "rasa_nlu.json"), format="json")
+    ds = from_rasa(os.path.join(FIXTURE_PATH, "rasa_nlu.json"), format="json")
     assert len(ds) == 5
     assert ds.texts[1] == "testing umläuts"
     assert ds.intents[0] == "restaurant_search"
@@ -47,14 +41,9 @@ def test_from_rasa_json():
 
 
 def test_from_rasa3_yml():
-    from rasa import __version__ as rasa_version
+    from nlubridge import from_rasa
 
-    if version.parse(rasa_version) < version.parse("3.0.0"):
-        return  # see 'test_from_rasa_json' for Rasa 2.x
-
-    from nlubridge.vendors.rasa3 import load_data
-
-    ds = load_data(os.path.join(FIXTURE_PATH, "rasa3_nlu.yml"))
+    ds = from_rasa(os.path.join(FIXTURE_PATH, "rasa3_nlu.yml"))
     assert len(ds) == 5
     assert ds.texts[1] == "testing umläuts"
     assert ds.intents[0] == "restaurant_search"
@@ -72,9 +61,9 @@ def test_from_rasa3_yml():
 
 
 def test_from_luis():
-    from nlubridge.vendors.luis import load_data
+    from nlubridge import from_luis
 
-    ds = load_data(os.path.join(FIXTURE_PATH, "luis_nlu.json"))
+    ds = from_luis(os.path.join(FIXTURE_PATH, "luis_nlu.json"))
     assert len(ds) == 5
     assert ds.texts[1] == "testing umläuts"
     assert ds.intents[0] == "restaurant_search"
